+++ php/Terminus/Commands/SiteCommand.php
@@ -511,7 +511,7 @@ public function backups($args, $assoc_args) {
           $target = sprintf('%s/%s', $target, $filename);
         }
         $this->log()->info('Downloading ... please wait ...');
-        if ($this->download($url->url, $target)) {
+        if (Request::download($url->url, $target)) {
           $this->log()->info('Downloaded {target}', compact('target'));
           return $target;
         } else {